feat:(config) Service configuration, easy integration to tkeel

+++ pkg/apirouter/tenant/v1/register.go
@@ -13,31 +13,42 @@ limitations under the License.
 package v1
 
 import (
-	"github.com/emicklei/go-restful"
-	restfulspec "github.com/emicklei/go-restful-openapi"
+	"net/http"
+
 	"github.com/tkeel-io/security/pkg/constants"
 	"github.com/tkeel-io/security/pkg/errcode"
 	"github.com/tkeel-io/security/pkg/models/dao"
-	"net/http"
+
+	"github.com/emicklei/go-restful"
+	restfulspec "github.com/emicklei/go-restful-openapi"
 )
 
 func AddToRestContainer(c *restful.Container) error {
+	var webservice *restful.WebService
+	for _, v := range c.RegisteredWebServices() {
+		if v.RootPath() == "v1" {
+			webservice = v
+			break
+		}
+	}
+	if webservice == nil {
+		webservice = &restful.WebService{}
+		webservice.Path("v1").
+			Produces(restful.MIME_JSON)
 
-	webservice := &restful.WebService{}
-	webservice.Path("/tenant").
-		Consumes(restful.MIME_JSON).
-		Produces(restful.MIME_JSON)
+		c.Add(webservice)
+	}
 
 	handler := newTenantHandler()
 
-	webservice.Route(webservice.POST("/").
+	webservice.Route(webservice.POST("tenant").
 		To(handler.Create).
 		Doc(" Create a tenant").
 		Reads(TenantCreteIn{}).
 		Returns(http.StatusOK, errcode.ErrMsgOK, TenantCreateOut{}).
 		Metadata(restfulspec.KeyOpenAPITags, []string{constants.APITagTenant}))
 
-	webservice.Route(webservice.GET("/").
+	webservice.Route(webservice.GET("tenants").
 		To(handler.Query).
 		Doc("get tenants").
 		Param(webservice.QueryParameter("tenant_id", "").Required(false)).
@@ -45,33 +56,32 @@ func AddToRestContainer(c *restful.Container) error {
 		Returns(http.StatusOK, errcode.ErrMsgOK, []*dao.Tenant{}).
 		Metadata(restfulspec.KeyOpenAPITags, []string{constants.APITagTenant}))
 
-	webservice.Route(webservice.DELETE("/{tenant_id}").
+	webservice.Route(webservice.DELETE("tenant/{tenant_id}").
 		To(handler.Delete).
 		Doc("delete a tenant").
 		Param(webservice.PathParameter("tenant_id", "tenant's ID").Required(true)).
 		Returns(http.StatusOK, errcode.ErrMsgOK, nil).
 		Metadata(restfulspec.KeyOpenAPITags, []string{constants.APITagTenant}))
 
-	webservice.Route(webservice.POST("/users").
+	webservice.Route(webservice.POST("tenant/users").
 		To(handler.UserCreate).
 		Doc("create a user").
 		Reads(UserCreateIn{}).
 		Returns(http.StatusOK, errcode.ErrMsgOK, dao.User{}).
 		Metadata(restfulspec.KeyOpenAPITags, []string{constants.APITagTenant}))
 
-	webservice.Route(webservice.GET("/users").
+	webservice.Route(webservice.GET("tenant/users").
 		To(handler.UserQuery).
 		Doc("get users").
 		Returns(http.StatusOK, errcode.ErrMsgOK, []dao.User{}).
 		Metadata(restfulspec.KeyOpenAPITags, []string{constants.APITagTenant}))
 
-	webservice.Route(webservice.DELETE("/users/{user_id}").
+	webservice.Route(webservice.DELETE("tenant/users/{user_id}").
 		To(handler.UserDelete).
 		Doc("delete a  users").
 		Param(webservice.PathParameter("user_id", "").Required(true)).
 		Returns(http.StatusOK, errcode.ErrMsgOK, nil).
 		Metadata(restfulspec.KeyOpenAPITags, []string{constants.APITagTenant}))
 
-	c.Add(webservice)
 	return nil
 }
